# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePath healthcare discovery, planning, appointments, demo tokens, reviews, and administration API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  hospitalId: number;
  name: string;
  specialtyId: number;
  specialty: string;
  qualifications: string;
  yearsExperience: number;
  available: boolean;
}